prueba 7.1

+++ instalacionProgramas.sh
@@ -6,12 +6,21 @@ sudo apt install curl -y
 sudo apt install docker.io -y
 sudo systemctl start docker
 sudo systemctl enable docker
-# sudo apt install vlc -y
+sudo apt install curl -y
+
+if curl -s --head https://get.helm.sh | grep "200 OK" > /dev/null; then
+    curl -fsSL -o get_helm.sh https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3
+    chmod 700 get_helm.sh
+    ./get_helm.sh
+else 
+    echo "Error: no se pudo acceder a internet para instalar helm"
+    exit 1
+fi
+
+
 curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
 sudo install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
-curl -fsSL -o get_helm.sh https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3
-chmod 700 get_helm.sh
-./get_helm.sh
+
 
 # Instalar eksctl
 curl --silent --location "https://github.com/weaveworks/eksctl/releases/latest/download/eksctl_$(uname -s)_amd64.tar.gz" | tar xz -C /tmp
@@ -20,13 +29,13 @@ sudo mv /tmp/eksctl /usr/local/bin
 
 # Crear cluster EKS
 eksctl create cluster \
---name eks-pin-final-1 \
+--name eks-mundos-e \
 --region us-east-1 \
 --node-type t3.small \
 --nodes 3 \
 --with-oidc \
 --ssh-access \
---ssh-public-key pin \
+--ssh-public-key cluster-key-* \
 --managed \
 --full-ecr-access \
 --zones us-east-1a,us-east-1b,us-east-1c
